requirements.txt
arrow==0.13.1
asn1crypto==0.24.0
beautifulsoup4==4.7.1
binaryornot==0.4.4
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
cookiejar==0.0.2
cryptography==2.6.1
future==0.17.1
html5lib==1.0.1
idna==2.8
Jinja2==2.10.1
jinja2-time==0.2.0
MarkupSafe==1.1.1
mechanize==0.4.2
mysql-connector-python==8.0.16
pager==3.3
poyo==0.4.2
protobuf==3.7.1
pyaes==1.6.1
pyasn1==0.4.5
pycparser==2.19
python-dateutil==2.8.0
python-telegram-bot==12.0.0b1
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.9.1
Telethon==1.7.7
tornado==6.0.2
urllib3==1.24.3
webencodings==0.5.1
whichcraft==0.5.2
